Support multiple ServicePorts per Service (#22)
This change is basically a global rewrite of the LoadBalancer-related
section of the CCM in order to natively support the typical use case
of multiple k8s `ServicePorts` defined in a `Service`, e.g. for an
Ingress Controller that exposes both ports 80 and 443 for the same
service.

Also, this change replaces most of the shellscripting related to
setting up then tearing down the cluster infrastructure required for the
integration tests by a Terraform configuration, resulting in a much more
manageable and reliable setup.

Finally, the documentation has been updated according to the changes
described above.